Authorization finish commit

+++ src/main/java/org/example/dto/JwtDTO.java
@@ -7,17 +7,17 @@ import org.example.enums.ProfileRole;
 @Getter
 @Setter
 public class JwtDTO {
-    private Long id;
+    private Integer id;
     private String username;
     private ProfileRole role;
 
-    public JwtDTO(Long id, String userName, ProfileRole role) {
+    public JwtDTO(Integer id, String userName, ProfileRole role) {
         this.id = id;
         this.username = userName;
         this.role = role;
     }
 
-    public JwtDTO(Long id) {
+    public JwtDTO(Integer id) {
         this.id = id;
     }
 }
